--- src/main/java/timf/voc/task/dto/DeliveryDriverDto.java
@@ -1,31 +1,0 @@
-package timf.voc.task.dto;
-
-import java.util.List;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import timf.voc.task.config.converter.EntityAndDtoConverter;
-import timf.voc.task.entity.DeliveryDriver;
-
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class DeliveryDriverDto {
-
-	private Long id;
-	private String name;
-	private Long salary;
-	private Long pendingPenaltyAmount;
-	private boolean isPenaltyDeducted;
-	private TransportCompanyDto transportCompany;
-	private List<VocDto> vocs;
-
-	public static DeliveryDriverDto from(DeliveryDriver deliveryDriver) {
-		return EntityAndDtoConverter.convert(deliveryDriver, DeliveryDriverDto.class);
-	}
-}
