--- src/main/java/timf/voc/task/dto/request/ClaimRequest.java
@@ -1,18 +1,0 @@
-package timf.voc.task.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class ClaimRequest {
-
-	private String personName;
-	private String content;
-}
