+++ src/main/java/timf/voc/task/interfaces/api/DeliveryDriverApiController.java
@@ -6,14 +6,14 @@
 //
 // import lombok.RequiredArgsConstructor;
 // import timf.voc.task.dto.api.ResponseDTO;
-// import timf.voc.task.dto.response.DeliveryDriverMyPageResponse;
-// import timf.voc.task.service.TransportCompanyService;
+// import timf.voc.task.interfaces.deliverydriver.DeliveryDriverMyPageResponse;
+// import timf.voc.task.domain.transportcompany.SimpleTransportCompanyService;
 //
 // @RestController
 // @RequiredArgsConstructor
 // public class DeliveryDriverApiController extends ApiController {
 //
-// 	private final TransportCompanyService transportCompanyService;
+// 	private final SimpleTransportCompanyService transportCompanyService;
 //
 // 	@GetMapping("/delivery-driver/my-page")
 // 	public ResponseDTO<DeliveryDriverMyPageResponse> getMyPage(@RequestParam(name = "id") Long id) {
